+++ doc/releaseNotes.md
@@ -504,3 +504,6 @@ This file describes the main feature changes for each readout.exe released versi
 
 ## v2.15.2 - 08/12/2022
 - Added ROC link Id for some log messages, as provided from new superpage metadata link field from ROC library.
+
+## next version
+- Superpage internal metadata not stored at beginning of page but separately in heap, by default. Behavior can be adjusted with MemoryPool::headerReservedSpace.